Refactored frames in anticipation of overhauling serialization.
