+++ BWEnv/CMakeLists.txt
@@ -42,9 +42,10 @@ add_library(BWEnvObj OBJECT
   src/user_actions.cc
   src/utils.cc
   src/zmq_server.cc
-  ../replayer/frame.cpp
-  ../replayer/frame_diff.cpp
   ../replayer/frame_serialization.cpp
+  ../replayer/keyframe.cpp
+  ../replayer/keyframe_serialization.cpp
+  ../replayer/frame_diff.cpp
   ../replayer/frame_diff_serialization.cpp
 )
