+++ BWEnv/VisualStudio/BWEnv.vcxproj
@@ -211,9 +211,10 @@
     <ClCompile Include="$(SolutionDir)\..\src\main.cc">
       <AdditionalIncludeDirectories Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">$(BWAPI_DIR)/include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
     </ClCompile>
-    <ClCompile Include="..\..\replayer\frame.cpp" />
-    <ClCompile Include="..\..\replayer\frame_diff.cpp" />
     <ClCompile Include="..\..\replayer\frame_serialization.cpp" />
+    <ClCompile Include="..\..\replayer\keyframe.cpp" />
+    <ClCompile Include="..\..\replayer\keyframe_serialization.cpp" />    
+    <ClCompile Include="..\..\replayer\frame_diff.cpp" />
     <ClCompile Include="..\..\replayer\frame_diff_serialization.cpp" />
     <ClCompile Include="..\src\config_manager.cc" />
     <ClCompile Include="..\src\controller.cc" />
